requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12.0
django-crispy-forms>=1.14.0
crispy-bootstrap5>=0.6
django-allauth>=0.51.0
django-widget-tweaks>=1.4.12
django-phonenumber-field[phonenumbers]>=6.1.0
Pillow>=9.2.0
gunicorn>=20.1.0
whitenoise>=6.2.0
psycopg2-binary>=2.9.3
python-dateutil>=2.8.2
drf-yasg>=1.21.7
xlsxwriter>=3.0.3  
reportlab>=3.6.11 
twilio>=7.0.0  
torch>=2.0.1
numpy>=1.24.3
transformers>=4.30.2